Refactor: renaming for consistency.
